move: 팀 패키지 통일에 따라 구조 변경
