+++ app/src/main/java/com/teamtriad/forpets/data/source/network/AbandonmentPublic.kt
@@ -1,4 +1,4 @@
-package com.teamtriad.forpets.model
+package com.teamtriad.forpets.data.source.network
 
 data class AbandonmentPublic(
     val response: ApiResponse
