+++ app/src/main/java/com/teamtriad/forpets/data/source/network/ApiResponse.kt
@@ -1,4 +1,4 @@
-package com.teamtriad.forpets.model
+package com.teamtriad.forpets.data.source.network
 
 data class ApiResponse(
     val header: ResponseHeader,
