+++ app/src/main/java/com/teamtriad/forpets/data/source/network/ResponseBody.kt
@@ -1,4 +1,4 @@
-package com.teamtriad.forpets.model
+package com.teamtriad.forpets.data.source.network
 
 data class ResponseBody(
     val items: Items,
