+++ app/src/main/java/com/teamtriad/forpets/ui/MainActivity.kt
@@ -1,10 +1,11 @@
-package com.teamtriad.forpets
+package com.teamtriad.forpets.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
+import com.teamtriad.forpets.R
 import com.teamtriad.forpets.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
