+++ app/src/main/java/com/teamtriad/forpets/ui/chat/ChatFragment.kt
@@ -1,10 +1,11 @@
-package com.teamtriad.forpets
+package com.teamtriad.forpets.ui.chat
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.teamtriad.forpets.R
 
 class ChatFragment : Fragment() {
