+++ app/src/main/res/layout/fragment_chat.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ChatFragment">
+    tools:context=".ui.chat.ChatFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
