+++ app/src/main/res/layout/fragment_transport.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TransportFragment">
+    tools:context=".ui.transport.TransportFragment">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
